http: Updating OutgoingMessage to contain the request method
Re-factoring to address PR comment. The logic here gets somewhat
complicated with regards to the if-elseif block circa line 434.
The goal of this PR is to prevent head requests from having their
Content-Length or Content-Encoding headers cleared except when
necessary, which is when the response code is 204 or 304. In trying
to address comment, I'm balancing between a larger re-factor than
absolutely needed and somewhat ugly control logic.

Refs: https://github.com/nodejs/node/pull/34231#discussion_r464078157